pyproject.toml
[tool.poetry]
name = "genetic-programming"
version = "0.1.0"
description = ""
authors = ["Piotr Karaś <pepe500p@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
antlr4-python3-runtime = "^4.13.1"
anytree = "^2.12.0"
pydantic = "^2.4.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
fawltydeps = "^0.13.1"
flake8 = "^6.0.0"
flake8-bandit = "^4.1.1"
flake8-bugbear = "^23.2.13"
flake8-builtins = "^2.1.0"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.10.1"
flake8-debugger = "^4.1.2"
flake8-isort = "^6.0.0"
flake8-print = "^5.0.0"
flake8-pyproject = "^1.2.2"
flake8-pytest = "^1.4"
flake8-pytest-style = "^1.7.2"
ipython = "^8.14.0"
isort = {version = "^5.12.0", extras = ["colors"]}
mypy = "^1.0.1"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-mock = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.fawltydeps]
code = ["src"]
deps = ["pyproject.toml"]
ignore_unused = [
    # Dev dependencies. This list should contain ALL of them!
    "black",
    "fawltydeps",
    "flake8",
    "flake8-bandit",
    "flake8-bugbear",
    "flake8-builtins",
    "flake8-commas",
    "flake8-comprehensions",
    "flake8-debugger",
    "flake8-isort",
    "flake8-print",
    "flake8-pyproject",
    "flake8-pytest",
    "flake8-pytest-style",
    "ipython",
    "isort",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[tool.mypy]
exclude = "src/antlr"
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = false

[tool.isort]
line_length = 120
order_by_type = false
profile = "black"

[tool.flake8]
# S101 == Use of assert detected. The enclosed code will be removed when compiling to optimised byte code.
# W503 == line break before binary operator. Conflicts with W504 - line break after binary operator.
ignore = ['S101', 'W503']
max-line-length = 120
exclude = "src/antlr"

[[tool.mypy.overrides]]
module = "antlr.*"
ignore_errors = true

[tool.black]
line-length = 120

# pytest settings
[tool.pytest.ini_options]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--strict-markers -p no:warnings --cov=. --cov-fail-under=90 --cov-config=.coveragerc"
